requirements.txt
flask>=3.0
flask-socketio>=5.3
yt-dlp>=2024.1
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0
gevent>=24.2
python-engineio>=4.8
python-socketio>=5.10
